Display storage and bandwidth in MB and GB
1. Refactored the code a bit, there was repetition, so moved it to a single place and called the function in App.vue and AppDashboard.vue
2. Added logic for converting value to MB if it is less than 1023 MB else it'll be converted to GB
