+++ src/App.vue
@@ -10,19 +10,16 @@
 </template>
 
 <script>
-import { onBeforeMount, ref, watch } from "@vue/runtime-core";
-import { useRoute } from "vue-router";
-import { getConfig } from "./services/app-config.service";
+import { onBeforeMount, ref } from "@vue/runtime-core";
+import { getConfig } from "@/services/app-config.service";
 import { useStore } from "vuex";
-import { stringToBuffer } from "./utils/cryptoUtils";
-import { fetchAllApps, fetchApp } from "@/services/dashboard.service";
-import bytes from "bytes";
+import { stringToBuffer } from "@/utils/cryptoUtils";
+import { fetchAndStoreAppConfig } from "@/services/app-config.service";
 import FullScreenLoader from "@/components/FullScreenLoader.vue";
 
 export default {
   components: { FullScreenLoader },
   setup() {
-    const route = useRoute();
     const store = useStore();
     const loading = ref("");
     const loadingMessage = ref("");
@@ -48,80 +45,7 @@ export default {
       if (!store.getters.appName && store.getters.accessToken) {
         loading.value = true;
         loadingMessage.value = "Fetching app configuration...";
-
-        const apps = await fetchAllApps();
-        if (apps.data.length) {
-          store.dispatch("updateAppConfigurationStatus", true);
-          const currentApp = apps.data[0];
-          const appId = currentApp.ID;
-          store.dispatch("updateAppName", currentApp.name);
-          store.dispatch("updateAppId", currentApp.ID);
-
-          const env = store.getters.env;
-          const chainType = ["ethereum", "polygon", "binance"][
-            currentApp.chain
-          ];
-          store.dispatch(env + "/updateChainType", chainType);
-          const unlimitedBytes = 10995116277760;
-
-          if (currentApp.storage_limit < unlimitedBytes) {
-            const storage = bytes(currentApp.storage_limit, {
-              unitSeparator: " ",
-            });
-            const storageValues = storage.split(" ");
-            store.dispatch(env + "/updateStorage", {
-              value: storageValues[0],
-              unit: storageValues[1],
-              isUnlimited: false,
-            });
-          } else {
-            store.dispatch(env + "/updateStorage", {
-              value: "",
-              unit: "",
-              isUnlimited: true,
-            });
-          }
-
-          if (currentApp.bandwidth_limit < unlimitedBytes) {
-            const bandwidth = bytes(currentApp.bandwidth_limit, {
-              unitSeparator: " ",
-            });
-            const bandwidthValues = bandwidth.split(" ");
-            store.dispatch(env + "/updateBandwidth", {
-              value: bandwidthValues[0],
-              unit: bandwidthValues[1],
-              isUnlimited: false,
-            });
-          } else {
-            store.dispatch(env + "/updateBandwidth", {
-              value: "",
-              unit: "",
-              isUnlimited: true,
-            });
-          }
-
-          const appDetails = await fetchApp(appId);
-
-          if (appDetails.data.cred) {
-            store.dispatch(
-              env + "/updateAuthDetails",
-              appDetails.data.cred.map((el) => {
-                return {
-                  type: el.verifier,
-                  authType: el.verifier,
-                  verifier: el.verifier,
-                  clientId: el.clientId,
-                  clientSecret: el.clientSecret,
-                  redirectUrl: el.redirectUrl,
-                };
-              })
-            );
-          } else {
-            store.dispatch(env + "/updateAuthDetails", []);
-          }
-        } else {
-          store.dispatch("updateAppConfigurationStatus", false);
-        }
+        await fetchAndStoreAppConfig();
         loading.value = false;
       }
     });
